+++ app/(tabs)/(onboard)/_layout.tsx
@@ -4,7 +4,7 @@ import { SafeAreaView, StyleSheet, TouchableOpacity } from 'react-native';
 import BackIcon from '@/components/icons/BackIcon';
 import { View } from '@/components/ui/atoms'
 import useToastStore from '@/stores/toastStore';
-import Toast from '@/components/Toast';
+import {Toast} from "@/components/ui/molecules";
 
 export default function OnboardLayout() {
 	const toasts = useToastStore((state) => state.toasts);
