+++ components/ui/molecules/index.ts
@@ -2,3 +2,4 @@ export { default as TabBar } from './TabBar'
 export { default as TabItem } from './TabItem'
 export { default as Tag } from './Tag'
 export { default as NavBar } from './NavBar'
+export { default as Toast } from './Toast'
